governance deployment

+++ packages/augur-core/src/contracts/gov/Governance.sol
@@ -90,7 +90,7 @@ contract Governance {
         bool support;
 
         /// @notice The number of votes the voter had, which were cast
-        uint96 votes;
+        uint256 votes;
     }
 
     /// @notice Possible states that a proposal may be in
@@ -269,7 +269,7 @@ contract Governance {
         Proposal storage proposal = proposals[proposalId];
         Receipt storage receipt = proposal.receipts[voter];
         require(receipt.hasVoted == false, "GovernorAlpha::_castVote: voter already voted");
-        uint96 votes = gov.getPriorVotes(voter, proposal.startBlock);
+        uint256 votes = gov.getPriorVotes(voter, proposal.startBlock);
 
         if (support) {
             proposal.forVotes = add256(proposal.forVotes, votes);
