+++ packages/augur-core/src/contracts/gov/IGovToken.sol
@@ -2,8 +2,8 @@ pragma solidity 0.5.15;
 
 
 interface IGovToken {
-    function totalSupply() external view returns (uint96);
-    function getPriorVotes(address account, uint blockNumber) external view returns (uint96);
-    function mintAllowance(address account) external view returns (uint96);
-    function mint(address _target, uint96 _amount) external returns (bool);
+    function totalSupply() external view returns (uint256);
+    function getPriorVotes(address account, uint blockNumber) external view returns (uint256);
+    function mintAllowance(address account) external view returns (uint256);
+    function mint(address _target, uint256 _amount) external returns (bool);
 }
